refactor(daa): merge DaaTrafficFilter and DaaEncoding
